+++ src/io/mod.rs
@@ -3,6 +3,5 @@ pub(crate) use self::{
     transport::ClickhouseTransport,
 };
 
-mod read_to_end;
 pub(crate) mod transport;
 pub(crate) mod stream;
